+++ data-compression/src/main/java/com/redhat/iot/api/SensorData.java
@@ -59,7 +59,7 @@ import javax.xml.bind.annotation.*;
   @JacksonXmlProperty(localName = "value")
   @XmlElement(name="value")
 
-  private Long value = null;
+  private Float value = null;
 
   @JsonProperty("units")
   @JacksonXmlProperty(localName = "units")
@@ -138,7 +138,7 @@ import javax.xml.bind.annotation.*;
   public void setType(String type) {
     this.type = type;
   }
-  public SensorData value(Long value) {
+  public SensorData value(Float value) {
     this.value = value;
     return this;
   }
@@ -150,10 +150,10 @@ import javax.xml.bind.annotation.*;
   * @return value
   **/
   @ApiModelProperty(value = "")
-  public Long getValue() {
+  public Float getValue() {
     return value;
   }
-  public void setValue(Long value) {
+  public void setValue(Float value) {
     this.value = value;
   }
   public SensorData units(String units) {
